:hammer: refactor: Change import path
- types 를 interface 로 수정
