+++ types/user/user-history-response.interface.ts
@@ -1,4 +1,4 @@
-import { UserInfo } from "./user-info.type";
+import { UserInfo } from "./user-info.interface";
 
 export interface UserHistoryResponse {
   histories: {
