+++ types/user/user-info-response.interface.ts
@@ -1,4 +1,4 @@
-import { UserInfo } from "./user-info.type";
+import { UserInfo } from "./user-info.interface";
 
 export interface UserInfoResponse extends UserInfo {
   blockedUsers: number[];
